+++ pages/accounts/notifications/index.tsx
@@ -14,7 +14,7 @@ export default function NotificationsPage() {
         </title>
       </Head>
       <Layout>
-        <Wrapper className="p-2 lg:p-5">
+        <Wrapper className="p-2 md:p-5">
           <Notifications />
         </Wrapper>
       </Layout>