+++ src/components/Button/index.tsx
@@ -1,9 +1,11 @@
 import styles from "./Button.module.css";
+import { Loading } from "@components/Loading";
 import type { MouseEventHandler, ReactNode } from "react";
 
-type Props = {
+type ButtonProps = {
   type?: "button" | "submit";
   className?: string;
+  loadingClassName?: string;
   children?: ReactNode;
   loading?: boolean;
   disabled?: boolean;
@@ -13,44 +15,22 @@ type Props = {
 export function Button({
   type,
   className,
+  loadingClassName,
   children,
   loading,
   disabled,
   onClick,
-}: Props) {
+}: ButtonProps) {
+  const classes = `
+    button 
+    ${loading ? styles.loading : ""} 
+    ${disabled ? styles.disabled : ""} 
+    ${className || ""}
+  `;
+
   return (
-    <button
-      onClick={onClick}
-      type={type || "button"}
-      className={`${styles.button} ${loading ? styles.loading : ""} ${
-        disabled ? styles.disabled : ""
-      } ${className}`}
-    >
-      {loading ? (
-        // TODO: create component for loading icon
-        <svg
-          className="animate-spin h-5 w-5 text-current"
-          xmlns="http://www.w3.org/2000/svg"
-          fill="none"
-          viewBox="0 0 24 24"
-        >
-          <circle
-            className="opacity-25"
-            cx="12"
-            cy="12"
-            r="10"
-            stroke="currentColor"
-            strokeWidth="4"
-          ></circle>
-          <path
-            className="opacity-75"
-            fill="currentColor"
-            d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-          ></path>
-        </svg>
-      ) : (
-        children
-      )}
+    <button onClick={onClick} type={type || "button"} className={classes}>
+      {loading ? <Loading className={loadingClassName || "w-6"} /> : children}
     </button>
   );
 }