+++ src/features/Home/index.tsx
@@ -4,5 +4,5 @@ import { Posts } from "@components/Posts";
 export function Home() {
   const { posts } = useHome();
 
-  return <Posts posts={posts} className="p-2 lg:p-5" />;
+  return <Posts posts={posts} className="p-2 md:p-5" />;
 }